feat: retry add loan to the basket when basket cookie has expired
